Fixed image orientation in image_view
Added commands to invert the y-axis of the viewbox (which by default
points upwards, whereas in most images the y-axis points downwards) and
to tell the image_item that the shape of the data is row-major i.e.
similar to a numpy array where the number of columns is given first
in the shape. With these two commands, images are then displayed with
the correct orientation in the gui.