Modified DataTableModel.update_node with ndim >=2
Previously with ndim >=2, DataTableModel.update_node would set
self.row_count to shape[0] and self.column_count to shape[1]. This did
not match with the DimsTableModel which always took the LAST two axes. I
modified the DataTableModel.update_node function so that the number of
rows and columns matches the last two axes in shape. This now matches
with the behaviour in the DimsTableModel.