Added an ImageModel and an ImageView
For parity with the other models/views and to hopefully make it easier
to develop in the future, an ImageModel was added to .models and an
ImageView was added to .views. These inherit the QAbstractItemModel and
QAbstractItemView.

Also linked the scrollbar on the ImageView to the DimsView and DimsModel
and vice versa. The behaviour is now that if you move the scrollbar, the
DimsView will update, showing you the slice corresponding to the image
displayed. If you manually edit the slice, when you confirm the changes,
the scrollbar will move to the correct place and the image will update.