Slice persistence when changing tabs
The HDF5Widget now stores the current slice (self.dims_model.shape) for
each tab when it is active in self.tab_dims. On changing tabs, the slice
is then reset to the previous value. If a tab is closed, the corresponding
key value pair will be deleted from self.tab_dims.