Persistence of tree index when chaning tabs
The HDF5Widget now stores the current tree index for each tab. This
enables the behaviour that you can e.g. start on an image tab, switch to
the table tab and then switch to a different tree index, then when you
switch back to the image tab, the previous state of that tab is
retained.