Enabled a blank image on ndim < 2
On the Image tab, if you change to a node on the tree with ndim < 2, the
previous image was still shown even though ImageModel.image_view was
None. Added a small h5 file to the resources/images folder and to the
resources qrc and py files. The h5 file contains only one node with a
3x3 array of zeros. This is read in on initialising the ImageModel and
is set as the ImageModel.image_view unless node.ndim >= 2. This changes
the above behaviour and if the user navigates to a node with ndim < 2, a
blank image is displayed in the ImageView.

The scrollbar is now hidden if ndim < 3.