Enabled having multiple ImageView tabs
On calling HDF5Widget.add_image, the ImageView is now stored in a dict
HDF5Widget.image_views where the key is the id (location in memory) of
the widget and the value is the ImageView widget itself. The functions
handle_dims_data_changed, handle_selection_changed and
handle_tab_changed have now been updated accordingly. The behaviour is
now that the user can have several Image tabs open and these remember
the slice when switching between the tabs etc.