Ensured ImageView.scrollbar range is always correct
Added the command self.scrollbar.setRange to self.update_image in
ImageView. This ensures the scrollbar range is always correct.

The scrollbar currently always corresponds to the FIRST dimension of the
node. TODO: add extra scroll bars with ndim > 3 or make a more complex
view for the DimsTableModel which can handle higher dimensions and
specifying which axis to use for the scrollbar.