Changed from blank image to switching widget off
When navigating to a node with ndim < 2, the previous behaviour was to
load a blank image 3x3 pixels of zeros. This behaviour has now been
replaced by simply hiding the viewbox widget if ndim < 2. This is much
neater, removes the need for the blank image in resources and solves the
problem that x, y coordinates would still be returned on the blank image
even if there was no such data at that node!