Simplified overridden functions in ImageModel
The ImageModel is a subclass of QAbsractItemModel and you have to
override the functions parent() and index() (and others). Previously, I
found some code on StackOverflow as an example of how to do this but it
didn't work. I have now replaced this and parent() and index() now
simply return a new QModelIndex using ImageModel.createIndex(). As the
ImageView never calls these functions, I think it is unnecessary for
them to do anything else. Short tests confirmed that the program still
works.