Enabled support for PyQt6 and PySide6
QtPy handles almost all the issues with switching between PyQt5,
PySide5, PyQt6 and PySide6, except: 1) the compiled resource system
used by PyQt5 is no longer available in PyQt6, therefore we simply
define a platform-independent relative path from the __file__ attribute
of main.py to the folder resources/images/ and add this as a search
path. This enables the icons to be loaded.

2) in main.py QApplication.setAttribute(Qt.AA_EnableHighDpiScaling) and
Qt.AA_UseHighDpiPixmaps cause errors in PyQt6 and PySide6. These are
defaults in those bindings anyway so we only need to set them in PyQt5
and PySide5

3) In the DimsTableModel, the values in the table were not centred with
   PyQt6 and PySide6. A workaround has been added and now the
   application has the same appearance in all Qt bindings.