Changed to pyproject.toml and src layout
Adopted the pyproject.toml format for specifying project metadata
because setup.py and setup.cfg will be depreciated at some point in the
future.

Adopted the scr/hdf5view layout, partly for convention but mostly
because with an editable install, the gui-script entry point did not
work properly with the flat layout. This problem was solved with the src
layout.