Bug fix with window geometry
The first time a user uses the program, some defualts are set for the
size of the window when it appears. There was a bug in the definition,
because QRect expects 4 ints but two ints and two floats were being
passed. This has now fixed that.

Version bump to reflect the change.