Fixed bug with number of recent files
The number of recent file actions in the menu is limited to
MAX_RECENT_FILES but on opening a new file which was not in the
list of recent files, it was added to the recent file list and
if there were more than MAX_RECENT_FILES in the list, an IndexError
was caused.

Added a few lines to mainwindow.open_file to ensure that the
list of recent files is never longer than MAX_RECENT_FILES.
Any new files are added to the start of the list and the end
of the list is clipped if necessary. This solved the above
problem.