Viewing rgb and rgba images/tables, and string format
Set defaults so that an node with ndim >= 3 and shape[-1] in [3,4]
is recognised as an rgb or rgba image. These images can be viewed in
an ImageView tab. Stacks of rgb or rgba images are also handled and
you can scroll through the first dimension with the scrollbar. These
table view of these nodes also shows the rgb or rgba values in each
cell (pixel). These are of course, only defaults and the user can
change the slice freely to see a different view. Greyscale images
are rendered at any node with ndim => 2, shape[-2] and shape[-1] > 1
and not shape[-1] in [3,4].

Changed the rendering of strings in the DataTable so that instead
of seeing "b'This is a string'", we now see "This is a string".