Updated status message for rgb and rgba images
When the cursor is over the image on an Image tab, a message
is printed to the bottom of the main window giving the x and y
coordinates of the cursor. Previously only the value at (x,y)
of luminance images was also displayed. This update handles rgb and
rgba images as well.