Refactoring of models and views
The function update_node() is now called on the ImageModel
and PlotModel, respectively, by the functions
HDF5Widget.add_image() and HDF5Widget.add_plot(). This allows
the filtering to be done by the models and the PlotView and
ImageView have been correspondingly streamlined. The model
functions set_dims() have also been simplified aand a new
function, get_dims_from_str() has been added to models.py
and is called by set_dims() in several models.
The data() method must be implemented in subclassed
QAbstractItemModel but is not needed for the ImageModel
and PlotModel classes, and so only returns a default
value in those cases.
The DataTableModel has been updated to use self.data_view
for all tables/slices. The data() method has been updated
accordingly and now the tables can be sliced in any way and
the correct data are shown.