Added functions to check node size before loading
The functions calculate_memory_ratio() and check_node_size()
were added to views.HDF5Widget. When a new dataset is selected
in the tree, these functions determine a critical size of the
dataset and find the ratio of this value to the available memory
on the system. If this ratio is more than
HDF5Widget.memory_ratio_limit, a QMessageBox.warning appears,
giving the user the chance to either cancel loading the dataset
or continue.

This protects against loading files which are larger than memory.
This is a provisional solution until dynamic loading can be
implemented.