Replacing obsolete QVariant() with None
fixes #13

Importing QVariant resulted in an import error with PySide6.
According to the docs: https://doc.qt.io/qtforpython-6/considerations.html#qvariant
QVariant and its constructor function are obsolete and should not be
used in future code. Instead None can be returned as an invalid
QVariant. The two functions where this appears:
ImageModel.data() and PlotModel.data() must be implemented when
subclassing but are not actually needed, therefore returning
None should work. Tested with PyQt5, PyQt6 and PySide6 and the gui
works with all of them.